Split image file into image_processing, image_filtering and image_masking.
