+++ .gitignore
@@ -1,3 +1,9 @@
-/.vscode/
+/.idea/
 /.pytest_cache/
+/.vscode/
+
 /env/
+
+/app/__pycache__/
+app/routers/__pycache__/
+tests/__pycache__/
